Rename GeneralRelativity Solutions directory
