+++ src/PointwiseFunctions/AnalyticSolutions/CMakeLists.txt
@@ -2,7 +2,7 @@
 # See LICENSE.txt for details.
 
 add_subdirectory(Burgers)
-add_subdirectory(EinsteinSolutions)
+add_subdirectory(GeneralRelativity)
 add_subdirectory(GrMhd)
 add_subdirectory(NewtonianEuler)
 add_subdirectory(Poisson)
